backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
psycopg[binary]==3.2.13
pydantic==2.8.2
pydantic-settings==2.4.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt>=4.0.0
python-multipart==0.0.9
apscheduler==3.10.4
httpx==0.27.2
email-validator==2.2.0
pytest==8.3.3
google-auth==2.32.0
anthropic==0.18.1
websockets==12.0
# Production infrastructure
alembic==1.13.1
slowapi==0.1.9
structlog==24.1.0
redis==5.0.1
# Email & 2FA
aiosmtplib==3.0.1
jinja2==3.1.3
pyotp==2.9.0
qrcode[pil]==7.4.2
# Object storage
boto3==1.35.0
# Monitoring
sentry-sdk[fastapi]==1.40.0
